Fix final-review issues: HMR pool guard, empty-update guard, confidence precision
- Restore globalThis HMR singleton guard for the pg Pool in db.ts, matching
  the pattern used by the old Prisma client, to avoid leaking pools on
  Next.js dev-mode hot reloads.
- Guard db.job.update / db.slide.update against empty data objects, which
  would otherwise produce malformed SQL.
- Change Slide.confidence from REAL to DOUBLE PRECISION to match the
  original Prisma Float mapping; migrated both local Postgres databases.

+++ src/lib/schema.sql
@@ -17,7 +17,7 @@ CREATE TABLE IF NOT EXISTS "Slide" (
     "extractedText" TEXT NOT NULL,
     "slideIntent" TEXT,
     "recommendedSections" TEXT,
-    "confidence" REAL,
+    "confidence" DOUBLE PRECISION,
     "sections" TEXT,
     "status" TEXT NOT NULL DEFAULT 'pending',
     "error" TEXT,
